requirements.txt
pandas
ta
schedule
datetime
config
pprint
json
warnings
python-telegram-bot
ccxt
